+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Leave trailing break at the end of text node in paragraph.
+- Remove empty text node from ATX heading.
 
 ## 0.2.0